requirements.txt
openai>=1.0.0
mcp>=1.0.0
atlassian-python-api>=4.0.0
